Add check for the existence of ramdisk-busybox.img before running FWTS tests
There could be situations when Linux with file "Image" in root directory exists
on the file system with a lower number than the file system with ACS tests
(Linux is installed on eMMC, ACS tests - external storage).
So, the wrong Linux without ramdisk-busybox.img is booting and FWTS tests are not performed.

Signed-off-by: Nazarii Obukhovskyi <nazarii.obukhovskyi@globallogic.com>