common: entry build scripts for 32bit arm
Adds an entry build script for 32bit Arm target devices.

FWTS component is not built/embedded due to an issue the in the
build env. A script use pkg-config which get host resources while
it may be preferred to use cross-compilation toolchain resources.

The build issue reports a assertion failure:
    /usr/include/glib-2.0/glib/gtypes.h:463:3: note: in expansion of macro ‘G_STATIC_ASSERT’
       G_STATIC_ASSERT(sizeof (unsigned long long) == sizeof (guint64));
       ^~~~~~~~~~~~~~~

Signed-off-by: Christophe Priouzeau <christophe.priouzeau@foss.st.com>
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>