get_source.sh: more flexible bbr-acs download
The get_bbr_acs_src() function uses a direct git clone to retrieve the
bbr-acs repository. This works only with git branches or tags, not with
commit ids.

Rework the download method to run in two steps with clone + checkout, to
allow referencing commit ids directly with the $ARM_BBR_TAG variable.

This fixes sources download for arm 32b.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>